Set up redux store. Wrote all thunk I think we will need based on api routes excluding wishlist. Renamed some api routes to agreed upon names like cart -> orderDetails
