+++ app/redux/actions.js
@@ -3,6 +3,7 @@ import {
     SET_USER,
     SET_SINGLE_PRODUCT,
     SET_ORDERS,
+    SET_ORDER_DETAILS,
 } from './constants';
 
 export const setProducts = products => {
@@ -30,5 +31,12 @@ export const setOrders = orders => {
     return {
         type: SET_ORDERS,
         orders,
+    }
+}
+
+export const setOrderDetails = orderDetails => {
+    return {
+        type: SET_ORDER_DETAILS,
+        orderDetails,
     }
 }
