+++ app/redux/index.js
@@ -1,9 +1,11 @@
 import { combineReducers } from 'redux';
+
 import {
     products,
     singleProduct,
     user,
     orders,
+    orderDetails,
 } from './reducers';
 
 export default combineReducers({
@@ -11,4 +13,5 @@ export default combineReducers({
     singleProduct,
     user,
     orders,
+    orderDetails,
 });
